Fix: Cannot read property 'Name' of undefined
When a Certificate with DNS validation is created using ACM, it normally
takes a few seconds before the validation DNS resource records are
generated and associated with the Certificate.  Therefore, the result of
the ACM describeCertificate API call immediately following the
Certificate creation would likely yield no ResourceRecord property (it's
undefined).  That causes the serverless command to fail with:

  Type Error ---------------------------------------------

  Cannot read property 'Name' of undefined

This fixes that issue by calling the ACM API every 2 seconds until the
ResourceRecords are found (for up to 1 minute) before continuing to the
rest of the code that depends on that data.